StockTradingPrototype-master/requirements.txt
aiohttp==3.7.4.post0
alpaca-trade-api==1.2.1
async-timeout==3.0.1
asyncpg==0.22.0
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
gevent==21.1.2
greenlet==1.1.0
idna==2.10
msgpack==1.0.2
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
polygon-api-client==0.1.11
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.16.0
twilio==6.59.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
websocket==0.2.1
websocket-client==0.59.0
websockets==8.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
